backend/requirements.txt
Django==3.2.16
djangorestframework==3.12.4
django-filter==2.4.0
djoser==2.3.0.0
python-dotenv==1.1.0
gunicorn==20.1.0
flake8==6.0.0 
flake8-isort==6.0.0

drf-extra-fields==3.7.0 
filetype==1.2.0
Pillow==11.2.1